feat/#21: 서버 컴포넌트 분리
