+++ app/diary/_component/DiaryClient.tsx
@@ -1,0 +1,102 @@
+"use client";
+
+import { useRouter } from "next/navigation";
+import Image from "next/image";
+import { useState } from "react";
+import Header from "@/components/Header";
+import BottomNaviation from "@/components/BottomNaviation";
+import TabNavigation from "@/app/diary/_component/TabNavigation";
+import DiaryCard from "@/app/diary/_component/DiaryCard";
+import MindLog from "@/app/diary/_component/MindLog";
+import MonthResolution from "@/app/diary/_component/MonthResolution";
+import { WeeklyEmotionItem } from "@/types/emotion.dto";
+
+interface DiaryClientProps {
+  weekRange: string;
+  weekData: WeeklyEmotionItem[];
+}
+
+export default function DiaryClient({ weekRange, weekData }: DiaryClientProps) {
+  const router = useRouter();
+  const [activeTab, setActiveTab] = useState<
+    "recent" | "emotion" | "medication"
+  >("recent");
+
+  return (
+    <div className="flex flex-col min-h-screen">
+      <Header
+        title={"기록"}
+        subHeader={
+          <TabNavigation activeTab={activeTab} onTabChange={setActiveTab} />
+        }
+      />
+
+      <div className="flex-1 overflow-y-auto p-4 pb-20 mt-28">
+        <MonthResolution />
+
+        <div className="flex justify-between mb-4">
+          <DiaryCard
+            type="emotion"
+            title={`오늘의 기분을\n기록하고 싶다면?`}
+            description="> 챗인지와 함께 기록하기"
+          />
+          <DiaryCard
+            type="medication"
+            title={`하루 한 번,\n오늘의 약 점검!`}
+            description="> 챗인지와 함께 기록하기"
+          />
+        </div>
+
+        <div className="mb-8">
+          <div className="mx-2 text-sm flex items-center gap-1">
+            <div>{weekRange}</div>
+            <Image
+              src="/ArrowRight.svg"
+              alt="full diary"
+              width={8}
+              height={10}
+              className="w-[8px] h-[10px] align-middle"
+            />
+          </div>
+          <div className="emoji-container mx-auto flex justify-between items-center mt-2">
+            {weekData.map(({ date, emotion, fullDate, dayOfWeek }) => (
+              <button
+                key={fullDate}
+                onClick={() => {
+                  if (emotion === null) {
+                    router.push(`/diary/${fullDate}/create`);
+                  } else {
+                    router.push(`/diary/${fullDate}`);
+                  }
+                }}
+                className={`emoji-container-item`}
+              >
+                <div className="flex flex-col items-center">
+                  <span className="text-sm my-1">{dayOfWeek}</span>
+                  {!emotion ? (
+                    <div className="w-[35px] h-[35px] bg-[#D9D9D9] rounded-[10px] text-2xl flex items-center justify-center font-extrabold text-gray-600">
+                      <div>+</div>
+                    </div>
+                  ) : (
+                    <Image
+                      src={`/Happy.png`}
+                      alt="{happy}"
+                      width={40}
+                      height={40}
+                      className="w-[40px] h-[34px]"
+                    />
+                  )}
+
+                  <span className="text-[10px] mt-1 text-gray-500">{date}</span>
+                </div>
+              </button>
+            ))}
+          </div>
+        </div>
+
+        <MindLog />
+      </div>
+      <BottomNaviation />
+    </div>
+  );
+}
