+++ package.json
@@ -7,10 +7,12 @@
     "build": "npx prisma generate && next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "jest"
+    "test": "jest",
+    "analyze": "ANALYZE=true pnpm run build"
   },
   "dependencies": {
     "@hookform/resolvers": "^5.0.1",
+    "@next/bundle-analyzer": "^15.3.1",
     "@prisma/client": "^6.5.0",
     "@radix-ui/react-label": "^2.1.2",
     "@radix-ui/react-slot": "^1.1.2",
