+++ types/emotion.dto.ts
@@ -1,6 +1,7 @@
 export interface WeeklyEmotionItem {
-  day: string; // ex: '월'
+  dayOfWeek: string; // ex: '월'
   date: string; // ex: '04.29'
+  fullDate: string; // ex: '2024-04-29'
   emotion: string | null;
 }
 
@@ -9,5 +10,5 @@ export interface WeeklyEmotionSummary {
     sunday: string; // ex: '04.27'
     saturday: string; // ex: '05.03'
   };
-  summary: WeeklyEmotionItem[];
+  weekData: WeeklyEmotionItem[];
 }
